refactor(connect-common): rename `core` group to `internal`
