+++ packages/connect-common/src/types/api/internal/removeAllListeners.ts
