+++ packages/connect-common/src/types/api/internal/uiResponse.ts
