+++ packages/connect-common/src/types/api/internal/updateConnectSettings.ts
